+++ example/android/app/src/main/kotlin/ir/tapsell/plus/tapsellplus_flutter_example/admob/NativeAdType.kt
@@ -1,0 +1,6 @@
+package ir.tapsell.plus.tapsellplus_flutter_example.admob
+
+enum class NativeAdType {
+    FULL,
+    LIST_ITEM
+}